# swagger.json
{"openapi":"3.0.0","paths":{"/tanks/all":{"get":{"operationId":"TankController_findAll","summary":"Get all tanks","description":"Returns a list of all tanks.","parameters":[],"responses":{"200":{"description":"Successful operation","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TankResponse"}}}}}},"tags":["Tanks"]}},"/tanks/tank":{"post":{"operationId":"TankController_createTank","summary":"Create a new tank","description":"Creates a new tank.","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTankDto"}}}},"responses":{"201":{"description":"Successfully created"},"400":{"description":"Bad request"}},"tags":["Tanks"]}},"/tanks/search":{"get":{"operationId":"TankController_search","summary":"Search for tanks","description":"Searches for tanks based on the provided query.","parameters":[{"name":"index","required":true,"in":"query","schema":{"type":"string"}},{"name":"name","required":true,"in":"query","schema":{"type":"string"}},{"name":"search","required":true,"in":"query","description":"Search query","schema":{}}],"responses":{"200":{"description":"Successful operation","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Tank"}}}}}},"tags":["Tanks"]}}},"info":{"title":"Tank Ops","description":"API description","version":"1.0","contact":{}},"tags":[{"name":"tanks","description":""}],"servers":[],"components":{"schemas":{"TankResponse":{"type":"object","properties":{}},"CreateTankDto":{"type":"object","properties":{"name":{"type":"string","description":"The name of the tank","example":"Tank 1"},"calibre":{"type":"number","description":"The calibre of the tank","example":100},"type":{"type":"number","description":"The type of the tank","enum":[0,1,2],"example":2},"weakSpots":{"type":"number","description":"The weak spots of the tank","example":4}},"required":["name","calibre","type","weakSpots"]},"Tank":{"type":"object","properties":{}}}}}